Pass table to AllocatorsTableFilters
